requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
numpy==1.22.0
pandas==1.3.3
Pillow==9.0.0
opencv-python==4.5.5.62
requests==2.27.1
python-dotenv==0.19.2
pymongo==4.0.1
pytest==7.0.0
supabase==2.0.0
flask-login==0.6.2 